+++ test/checkout/faker.py
@@ -2,6 +2,7 @@ import decimal
 
 import pydantic
 
+from checkout.gateway import adapter
 from checkout.standard_types import money
 from checkout.gateway import service
 
@@ -38,5 +39,39 @@ class PaymentResponseFaker:
     def with_rejected_transaction() -> service.PaymentResponse:
         return service.PaymentResponse(
             approval_number="",
-            status="APPROVED"
+            status="REJECTED"
+        )
+
+
+class StubApprovedTransactionCardNotPresentProvider(adapter.CardNotPresentProvider):
+    def __init__(self,
+                 approval_number: str = "",
+                 network: str = "CBK") -> None:
+        self.approval_number = approval_number
+        self.network = network
+
+    def sale(self, transaction: adapter.Transaction) -> adapter.TransactionResponse:
+        return adapter.TransactionResponse(
+            transaction_id="fake-transaction-id",
+            network=self.network,
+            response_code="00",
+            response_message="Approved or completed successfully",
+            status=adapter.TransactionStatus.APPROVED,
+            approval_number=self.approval_number,
+        )
+
+
+class StubRejectedTransactionCardNotPresentProvider(adapter.CardNotPresentProvider):
+    def __init__(self,
+                 network: str = "CBK") -> None:
+        self.network = network
+
+    def sale(self, transaction: adapter.Transaction) -> adapter.TransactionResponse:
+        return adapter.TransactionResponse(
+            transaction_id="fake-transaction-id",
+            network=self.network,
+            response_code="05",
+            response_message="Do not honor",
+            status=adapter.TransactionStatus.REJECTED,
+            approval_number="",
         )
